Add files via upload
Unzip and run ps2exposed.exe then use the buttons to configure each path. important if pcsx2 is running this exe will assume you are trying to run a game and the GUI will not open. 

After configuration point your frontend to the ps2exposed.exe insted of the pcsx2.exe emulator.